+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Bug #12: Update value default `override` `true` to `false` in `HasLinkItemActiveClass` (@terabytesoftw)
 - Enh #13: Add trait `HasLinkItemDisableClass` class (@terabytesoftw)
 - Bug #14: Update value default `override` `true` to `false` in `HasLinkActiveClass` and `HasLinkDisableClass` (@terabytesoftw)
+- Enh #15: Add trait `HasTemplateItems` class (@terabytesoftw)
 
 ## 0.1.0 March 5, 2024